stbt.Keyboard: Increase navigate_timeout to 60
I can imagine a Page Object with a slow `selection` implementation, plus
`verify_every_keypress=True`, plus a missed keypress or two, could
easily exceed this timeout.

Its purpose is to avoid infinite loops. If a test takes a bit longer to
fail when there's a real bug, that's an acceptable trade-off if we get
less false failures.