stbt.Keyboard: Use stable frame for page.refresh()
I was testing the Search keyboard on a Comcast Xfinity box. It had
pressed KEY_RIGHT several times and the selection had settled on one of
the intermediate keys for long enough that it tricked our
`stable_secs=0.5` --see the previous commit's message for more details
about this set-top-box-- but when we called `page.refresh()` on the next
line, the new frame of video had started moving again, giving an invalid
region which caused `find_key` to raise this:

> ValueError: Query {'region': Region(...)} doesn't match any key in the
> keyboard

This commit also fixes the assert -- it no longer asserts on
TransitionStatus.START_TIMEOUT so our missed-keypress retries (added a
few commits ago) will actually work now.