MotionDiff: Take color information into account
Previously, MotionDiff converts the 2 frames to grayscale before
comparing them. This means that it doesn't catch text in 2 different
colors, but with similar overall intensity. The new algorithm does a
simple color distance calculation (consistent with
`find_selection_from_background` and with OCR's `text_color` parameter)
before doing the threshold.

Along with this, the parameter name has changed from `noise_threshold`
to `threshold` and its range has changed from [1.0, 0.0] to [0, 255].

I have copied the old implementation as `GrayscaleDiff`. This allows
existing users to upgrade to v33 while still keeping the old behaviour,
by running the following code early, for example at the top level of
`tests/__init__.py` (this is similar to what we did for `press_and_wait`
when we changed its algorithm in v32):

    stbt.detect_motion.differ = stbt.GrayscaleDiff